Define a DTO file
The following creates a transport.go file for all common DTOs. This
should make it clear about where the commonality aligns for each
request.
