+++ charmhub/client.go
@@ -99,11 +99,10 @@ func NewClient(config Config) (*Client, error) {
 		return nil, errors.Annotate(err, "constructing find path")
 	}
 
-	var (
-		httpClient   = DefaultHTTPTransport()
-		apiRequester = NewAPIRequester(httpClient)
-		restClient   = NewHTTPRESTClient(apiRequester)
-	)
+	httpClient := DefaultHTTPTransport()
+	apiRequester := NewAPIRequester(httpClient)
+	restClient := NewHTTPRESTClient(apiRequester)
+
 	return &Client{
 		infoClient: NewInfoClient(infoPath, restClient),
 		findClient: NewFindClient(findPath, restClient),
@@ -118,61 +117,4 @@ func (c *Client) Info(ctx context.Context, name string) (InfoResponse, error) {
 // Find searches for a given charm for a given name from charmhub.
 func (c *Client) Find(ctx context.Context, name string) ([]FindResponse, error) {
 	return c.findClient.Find(ctx, name)
-}
-
-type ChannelMap struct {
-	Channel  Channel  `json:"channel,omitempty"`
-	Revision Revision `json:"revision,omitempty"`
-}
-
-type Channel struct {
-	Name       string   `json:"name"`
-	Platform   Platform `json:"platform"`
-	ReleasedAt string   `json:"released-at"`
-	Risk       string   `json:"risk"`
-	Track      string   `json:"track"`
-}
-
-type Platform struct {
-	Architecture string `json:"architecture"`
-	OS           string `json:"os"`
-	Series       string `json:"series"`
-}
-
-type Revision struct {
-	ConfigYAML   string     `json:"config-yaml"`
-	CreatedAt    string     `json:"created-at"`
-	Download     Download   `json:"download"`
-	MetadataYAML string     `json:"metadata-yaml"`
-	Platforms    []Platform `json:"platforms"`
-	Revision     int        `json:"revision"`
-	Version      string     `json:"version"`
-}
-
-type Download struct {
-	HashSHA265 string `json:"hash-sha-265"`
-	Size       int    `json:"size"`
-	URL        string `json:"url"`
-}
-
-type Charm struct {
-	Categories  []Category        `json:"categories"`
-	Description string            `json:"description"`
-	License     string            `json:"license"`
-	Media       []Media           `json:"media"`
-	Publisher   map[string]string `json:"publisher"`
-	Summary     string            `json:"summary"`
-	UsedBy      []string          `json:"used-by"`
-}
-
-type Category struct {
-	Featured bool   `json:"featured"`
-	Name     string `json:"name"`
-}
-
-type Media struct {
-	Type   string `json:"type"`
-	URL    string `json:"url"`
-	Width  int    `json:"width,omitempty"`
-	Height int    `json:"height,omitempty"`
 }
